refactor(store): split storage repositories
