+++ crates/bazel-mcp-store/src/lib.rs
@@ -1,10 +1,21 @@
 //! Database-free, crash-recoverable invocation evidence storage.
 
 mod cursor;
+mod deferred_repository;
 mod files;
+mod index;
+mod invocation_repository;
+mod manifest;
+mod query_paging;
+mod record;
+mod retention;
 mod storage;
 
 pub use files::InvocationPaths;
+pub use record::{
+    CoverageHeader, HydratedInvocation, InvocationDetails, InvocationHeader,
+    InvocationSummaryHeader,
+};
 pub use storage::{InvocationCompletion, Store, StoreError, StoreIoStats, StoreStartupStats};
 
 /// Parser entry point used by adversarial cursor tests and fuzzing.
